Move some util classes into Utils
