+++ unity/EditorPlugin/Utils/BoundedSynchronizedQueue.cs
@@ -3,18 +3,19 @@ using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 
-namespace JetBrains.Rider.Unity.Editor.NonUnity
+namespace JetBrains.Rider.Unity.Editor.Utils
 {
-  public class BoundedSynchronizedQueue<T> where T : class
+  internal class BoundedSynchronizedQueue<T> where T : class
   {
     private readonly int myMaxSize;
     private readonly Queue<T> myQueue;
     private readonly object myLockObject = new object();
+
     public BoundedSynchronizedQueue(int maxSize)
     {
-      if (maxSize < 0)         
+      if (maxSize < 0)
         throw new ArgumentOutOfRangeException(nameof (maxSize), "ArgumentOutOfRange_NeedNonNegNum");
-      
+
       myMaxSize = maxSize;
       myQueue = new Queue<T>();
     }
@@ -32,12 +33,12 @@ namespace JetBrains.Rider.Unity.Editor.NonUnity
     {
       if (input == null)
         throw new ArgumentNullException("input is null.");
-      
+
       lock (myLockObject)
       {
         myQueue.Enqueue(input);
         if (myQueue.Count >= myMaxSize)
-          myQueue.Dequeue(); // limit max size  
+          myQueue.Dequeue(); // limit max size
       }
     }
   }
