+++ unity/EditorPlugin/Utils/VersionExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace JetBrains.Rider.Unity.Editor.NonUnity
+namespace JetBrains.Rider.Unity.Editor.Utils
 {
-  public static class VersionExtensions
+  internal static class VersionExtensions
   {
     public static bool TryParse(string input, out Version version)
     {
@@ -10,17 +10,17 @@ namespace JetBrains.Rider.Unity.Editor.NonUnity
       {
         if (!input.Contains("."))
           input += ".0"; // new Version fails if case of single digit
-        
+
         version = new Version(input);
         return true;
       }
       catch (ArgumentException)
       {
-      } // can't put loggin here because ot fire on every symbol
+      } // can't put logging here because ot fire on every symbol
       catch (FormatException)
       {
       }
-      
+
       version = new Version(0,0);
 
       return false;
